This is the first release of a very easy to use mathematical parser for the swift language. This project is in its infancy so some features may not be fully functioning, however basic math operations should work just fine.
After searching for an easy to use mathematical parser and not being
able to find one I decided to make my own parser. This parser is very
simplistic, by that I mean that it does not use a grammar definition.
It uses the shunting yard algorithm at its core.
Use:
 Create an instance of MathmaticalTree
call initWithFormula(formula: String) // this will return the root node
of the tree.
Each node is of type ExpressionNode, which has a method
printNodeAndChildren(prefix: String, isTail: Bool) which will print the
node and all of the tree below it.